+++ src/api/api.ts
@@ -1,19 +1,18 @@
-import { chooseOneSurahFromSet, generateRandomSurahs } from "../utils/surahUtils.ts";
 import axios from "axios";
 
-export interface SurahData {
+interface SurahData {
     id: number;
     arabicName: string;
     englishName: string;
     englishTranslationName: string;
     numberOfAyahs: number;
     audioUrl: string | null;
-    chosenSurah: boolean
+    chosenSurah: boolean;
 }
 
 const API_BASE_URL = 'http://api.alquran.cloud/v1/surah/'
 
-async function fetchSurahData(surahNumber: number) {
+export async function fetchSurahData(surahNumber: number) {
     try {
         const response = await axios.get(`${API_BASE_URL}${surahNumber}/ar.alafasy`, {
             headers: { Accept: 'application/json' },
@@ -24,11 +23,10 @@ async function fetchSurahData(surahNumber: number) {
         console.error('Error fetching surah data:', error);
         throw error;
     }
-
 }
 
 
-async function getSurahDetails(surahNumber: number, isChosen: boolean): Promise<SurahData | null> {
+export async function getSurahDetails(surahNumber: number, isChosen: boolean): Promise<SurahData | null> {
     try {
         const data = await fetchSurahData(surahNumber)
 
@@ -51,21 +49,3 @@ async function getSurahDetails(surahNumber: number, isChosen: boolean): Promise<
 
 
 }
-
-async function generateSurahData(numberOfSurahs: number): Promise<SurahData[]> {
-    const randomSurahArray = Array.from(generateRandomSurahs(numberOfSurahs))
-    const randomlyChosenSurah = chooseOneSurahFromSet(randomSurahArray)
-
-    console.log('random', randomlyChosenSurah)
-    const surahDataList = await Promise.all(randomSurahArray.map((surahNumber) =>
-        getSurahDetails(surahNumber, surahNumber === randomlyChosenSurah)
-    ))
-
-
-    return surahDataList.filter((data): data is SurahData => data !== null);
-
-}
-
-generateSurahData(4).then((data) => {
-    console.log(data);
-});
